changer reporter basemodel to gpt4 & refactored agents to different files
