+++ backend/app/config/agents.yaml
@@ -46,6 +46,7 @@ job_role_researcher:
     Research in-depth details of the {job} role at {company}.
     Extract the official job description, day-to-day responsibilities, required skills,
     team dynamics, and the entire interview process including sample questions and candidate feedback.
+    Research and compile data on compensation, benefits, career growth opportunities, and overall workplace culture
   backstory: >
     You are an expert in job market analysis with a deep understanding of technical roles.
     Your task is to compile a detailed and structured overview of the {job} role that can be easily integrated into a table.
@@ -67,17 +68,17 @@ compensation_career_culture_researcher:
   llm: azure/gpt-4o
 
 
-# Agent for Data Sources & References Aggregation
-references_researcher:
+interview_preparation_researcher:
   role: >
-    {company} Research References Aggregator
+    {company} {job} Interview Preparation Researcher
   goal: >
-    Compile a consolidated and well-organized list of all sources and references used in the research of
-    {company} and the {job} role. Organize the references by section and ensure each source includes a direct URL.
+    Gather, synthesize, and organize real‑world interview experiences, common questions, sample answers, 
+    and actionable preparation advice for candidates applying to the {job} role at {company}. Structure output by topic (e.g., technical questions, behavioral questions, interview format), and include direct URLs to each source so users can verify and dive deeper.
   backstory: >
-    You are highly skilled in information synthesis and source verification.
-    Your role is to ensure all gathered data is properly referenced, providing credibility and traceability.
+    You’re an expert at mining first‑hand candidate accounts (Glassdoor, Blind, Reddit, LeetCode Discuss, Medium blogs, etc.), 
+    vetting source credibility, and distilling complex anecdotal data into concise, high‑impact prep guidance. Your mission is to empower job seekers with clear, step‑by‑step strategies to excel in every stage of the interview process for the {job} role at {company}.
   llm: azure/gpt-4o
+
 
 reporting_analyst:
   role: >
@@ -89,4 +90,4 @@ reporting_analyst:
     You are a meticulous analyst with a strong background in data reporting.
     Using the research data provided, your job is to format this information into a table,
     ensuring that all the relevant details about {company} and the {job} role are clearly presented.
-  llm: azure/gpt-4o
+  llm: azure/o3-mini
