+++ backend/app/services/clients/crew_client/agents/table_agent.py
@@ -1,0 +1,96 @@
+from crewai import Agent, Crew, Process, Task
+from crewai.project import CrewBase, agent, crew, task, before_kickoff, after_kickoff
+from config.logger import logger
+from services.clients.crew_client.crew_client_helpers import * 
+
+@CrewBase
+class TableMakerCrew():
+    @before_kickoff
+    def before_kickoff_function(self, inputs):
+        logger.info(f"Before kickoff function with inputs: {inputs}")
+        return inputs # You can return the inputs or modify them as needed
+    
+    @after_kickoff
+    def after_kickoff_function(self, result):
+        logger.info(f"After kickoff function with result: {result}")
+        return result # You can return the result or modify it as needed
+    
+    @agent
+    def company_corporate_researcher(self) -> Agent:
+        return create_agent_from_yaml("company_corporate_researcher", [Tools["perplexity_search_tool"]])
+    
+    @task
+    def company_corporate_research_task(self) -> Task:
+        return Task(
+            config=tasks_config['company_corporate_research_task'],
+            description=tasks_config['company_corporate_research_task']['description'],
+            expected_output=tasks_config['company_corporate_research_task']['expected_output'],
+            agent=self.company_corporate_researcher()
+        )
+    
+    
+    @agent
+    def mission_vision_researcher(self) -> Agent:
+        return create_agent_from_yaml("mission_vision_researcher", [Tools["perplexity_search_tool"]])
+    @task
+    def mission_vision_researcher_task(self) -> Task:
+        return Task(
+            description=tasks_config['mission_vision_research_task']['description'],
+            expected_output=tasks_config['mission_vision_research_task']['expected_output'],
+            agent=self.mission_vision_researcher()
+        )
+    @agent
+    def financial_market_researcher(self) -> Agent:
+        return create_agent_from_yaml("financial_market_researcher", [Tools["perplexity_search_tool"]])
+
+    @task
+    def financial_market_researcher_task(self) -> Task:
+        return Task(
+            config=tasks_config['financial_market_research_task'],
+            description=tasks_config['financial_market_research_task']['description'],
+            expected_output=tasks_config['financial_market_research_task']['expected_output'],
+            agent=self.financial_market_researcher()
+        )
+    @agent 
+    def job_role_researcher(self) -> Agent:
+        return create_agent_from_yaml("job_role_researcher", [Tools["perplexity_search_tool"]])
+    @task
+    def job_role_researcher_task(self) -> Task:
+        return Task(
+            config=tasks_config['job_role_research_task'],
+            description=tasks_config['job_role_research_task']['description'],
+            expected_output=tasks_config['job_role_research_task']['expected_output'],
+            agent=self.job_role_researcher()
+        )
+    @agent
+    def compensation_career_culture_researcher(self) -> Agent:
+        return create_agent_from_yaml("compensation_career_culture_researcher", [Tools["perplexity_search_tool"]])
+    @task
+    def compensation_career_culture_researcher_task(self) -> Task:
+        return Task(
+            config=tasks_config['compensation_career_culture_research_task'],
+            description=tasks_config['compensation_career_culture_research_task']['description'],
+            expected_output=tasks_config['compensation_career_culture_research_task']['expected_output'],
+            agent=self.compensation_career_culture_researcher()
+        )
+    @agent
+    def reporting_analyst(self) -> Agent:
+        return create_agent_from_yaml("reporting_analyst")
+    @task
+    def reporting_task(self) -> Task:
+        return Task(
+            description=tasks_config['reporting_task']['description'],
+            expected_output=tasks_config['reporting_task']['expected_output'],
+            agent=self.reporting_analyst(),
+            output_json=JobInformation,
+        )
+    @crew
+    def crew(self) -> Crew:
+        """Creates the LatestAiDevelopment crew"""
+        # print(self.agents)
+        return Crew(
+            agents=self.agents, # Automatically created by the @agent decorator
+            tasks=self.tasks, # Automatically created by the @task decorator
+            process=Process.sequential,
+            verbose=True,
+        )
